requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2
django-ckeditor==6.1.0
django-js-asset==1.2.2
django-svg-image-form-field==1.0.0
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
mysqlclient==2.0.3
django-allauth==0.45.0   
django-cleanup==5.2.0
django-unused-media==0.2.2
django-storages==1.11.1
boto3==1.18.28